feat: add new words for user to guess
Commit description:
I added a variety of additional words for users to guess, with a particular focus on country names. This update includes country names in both English and Portuguese, expanding the range of words available for users to interact with and guess correctly.

Changes made:

Added a collection of country names in English and Portuguese to the list of words for the user to guess.
Expanded the word selection to include a diverse range of country names, enhancing the gameplay experience.
Users can now encounter and guess country names in both English and Portuguese during gameplay.